Update beambreak_manager.ino with proper data structs for data and cleaner code
Used a queue (circular buffer under the hood) for storing the sensor pins and cleaned up to code to look prettier. Also defined a struct to represent each of the four little boxes that will hold a pair of beambreak sensors. The way the program loop works should be fine when adding more beam compartments in the mix

TODO:
- changed the value of DISTANCE to what it actually is
- implement some means of tracking which electromagnet to turn on, will probably be similar to the beam compartment struct that I've got going on